backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
snowflake-connector-python==3.7.1
snowflake-snowpark-python==1.11.1
sqllineage==1.4.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
structlog==23.2.0
prometheus-client==0.19.0
asyncio-mqtt==0.16.1
aioredis==2.0.1
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
celery==5.3.4
redis==5.0.1
boto3==1.34.0
python-dotenv==1.0.0
pytz==2023.3